- Move chunk tick fix to a feature
